font color debug
