+++ src/components/navbars/sidebar/display/fontColor.tsx
@@ -1,26 +1,16 @@
-import React, { useState } from 'react'
-
-import { RootState } from '../../../../store';
-import { useDispatch, useSelector } from 'react-redux';
 import Theme from '../../../theme';
-import {Button, ThemeProvider, FormatColorTextIcon } from '../../../../muiImports';
-import Testbutton from './testbutton';
+import { ThemeProvider } from '../../../../muiImports';
+import ColorWheelButton from './colorButton';
 
 
 export default function ShowFrequency() {
 
   const {myTheme} = Theme();
 
-  
   return (
     <div>
       <ThemeProvider theme={myTheme}>
-        {/* <Button id="demo-customized-button" variant="contained" disableElevation
-                // onClick={handleClick}
-                sx={{ width: 50, height: 30 }}>
-          {<FormatColorTextIcon />}
-        </Button> */}
-        <Testbutton />
+        <ColorWheelButton />
       </ThemeProvider>
     </div>
   );
